more server refactor
